Created participant list with a status.
Created two services to return a list of participants with a status.
Now we don't need to use 4 services to get accepted, rejected, pending and pendingGrouped. With this two services, the idea is to list ans clasyfy in frontend using status field.

function in service_utils is to clasify id fast.